Adds LslEventListener Class.
An instance of this class will listen to incoming LSL events. Per incoming LSL event it starts a child worker (thread) that will decode the message and and depending on the content will create a classifier instance, collect data, train or predict the models. Instances of this class are executed in it's own thread of control as well.